requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
python-dotenv
requests
werkzeug
gunicorn